+++ moor_generator/lib/src/analyzer/options.g.dart
@@ -16,7 +16,7 @@ MoorOptions _$MoorOptionsFromJson(Map<String, dynamic> json) {
       'use_data_class_name_for_companions',
       'use_column_name_as_json_key_when_defined_in_moor_file',
       'generate_connect_constructor',
-      'use_experimental_inference',
+      'legacy_type_inference',
       'sqlite_modules',
       'eagerly_load_dart_ast'
     ]);
@@ -44,9 +44,9 @@ MoorOptions _$MoorOptionsFromJson(Map<String, dynamic> json) {
       generateConnectConstructor: $checkedConvert(
               json, 'generate_connect_constructor', (v) => v as bool) ??
           false,
-      useExperimentalInference: $checkedConvert(
-              json, 'use_experimental_inference', (v) => v as bool) ??
-          false,
+      legacyTypeInference:
+          $checkedConvert(json, 'legacy_type_inference', (v) => v as bool) ??
+              false,
       eagerlyLoadDartAst:
           $checkedConvert(json, 'eagerly_load_dart_ast', (v) => v as bool) ??
               false,