+++ sqlparser/test/analysis/types2/misc_cases_test.dart
@@ -39,9 +39,7 @@ const Map<String, ResolvedType> _types = {
 };
 
 SqlEngine _spawnEngine() {
-  return SqlEngine(EngineOptions(enableExperimentalTypeInference: true))
-    ..registerTable(demoTable)
-    ..registerTable(anotherTable);
+  return SqlEngine()..registerTable(demoTable)..registerTable(anotherTable);
 }
 
 void main() {